refactor(folder-browser): G.87.5 — extract useFolderBrowserKeyboardNav hook
